Improved sequential wait
While running on limited resource system like that with a single Nvidia RTX3080Ti, requests to Ollama must be made sequentially which is not guaranteed even with sequential Python programming.
This commit improves the sequential wait in the code especially when interacting with the latest Ollama version (at this time).